requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pytesseract>=0.3.7
pillow>=8.3.1
spacy>=3.0.0
python-docx>=0.8.10
pdfplumber>=0.5.28
pdf2image>=1.14.0
requests>=2.26.0
pylint>=2.11.1




# Install these first (Ubuntu/Debian example):
# sudo apt-get install -y \
#     tesseract-ocr \
#     poppler-utils \
#     libtesseract-dev \
#     libleptonica-dev \
#     tesseract-ocr-eng \
#     tesseract-ocr-script-latn

# For Windows users, use:
# https://github.com/UB-Mannheim/tesseract/wiki
# and install poppler from: https://github.com/oschwartz10612/poppler-windows